Refactor the creation of the request
Put the creation of the request into one function.
